📝 docs(README.md, usage.md): Update Mattermost version and reorganize usage features (#134)
* 📝 docs(README.md, usage.md): Update Mattermost version and reorganize usage features
The README.md file was updated to specify that the Mattermost AI Plugin is compatible with Mattermost v9.6 and above. This change was made to ensure users are aware of the minimum version requirement for the plugin. The usage.md file was also updated to reorganize the features of the plugin. Some features were renamed and others were moved around to improve the flow and readability of the document. This was done to make the document more user-friendly and easier to understand.

* 📝 docs: Update plugin name from 'Mattermost AI Plugin' to 'Mattermost AI Copilot Plugin'
The plugin name has been updated across all documentation files to reflect the new name 'Mattermost AI Copilot Plugin'. This change was made to better represent the functionality of the plugin, as it acts as a copilot by providing AI capabilities within the Mattermost platform.

* 📑 docs: adjust v1.0 feature documentation
Fixes #132